Extract build plugin for instrumentation testing projects that aren't muzzled / published.
