+++ buildSrc/src/main/kotlin/otel.library-instrumentation.gradle.kts
@@ -1,8 +1,7 @@
 plugins {
-  id("otel.java-conventions")
+  id("otel.instrumentation-conventions")
   id("otel.jacoco-conventions")
   id("otel.publish-conventions")
-  id("otel.instrumentation-conventions")
 }
 
 extra["mavenGroupId"] = "io.opentelemetry.instrumentation"
