+++ testing-common/integration-tests/integration-tests.gradle
@@ -1,6 +1,6 @@
-ext.skipPublish = true
-
-apply plugin: "otel.javaagent-instrumentation"
+plugins {
+  id("otel.javaagent-testing")
+}
 
 dependencies {
   implementation project(':testing-common:library-for-integration-tests')
