+++ testing/armeria-shaded-for-testing/armeria-shaded-for-testing.gradle
@@ -8,7 +8,11 @@ apply plugin: "otel.java-conventions"
 apply plugin: "otel.publish-conventions"
 
 dependencies {
-  implementation "com.linecorp.armeria:armeria-junit5:1.8.0"
+  implementation("com.linecorp.armeria:armeria-junit5:1.8.0") {
+    // We don't use JSON features of Armeria but shading it in can cause version conflicts with
+    // instrumentation tests that do.
+    exclude group: "com.fasterxml.jackson.core"
+  }
 }
 
 shadowJar {
